Removed 'fs-extra' from website dependencies. Moved and renamed utility.js to fs-utility.js for future use.
