+++ src/package.json
@@ -28,7 +28,6 @@
     "debug": "^3.1.0",
     "express": "^4.16.3",
     "express-session": "^1.15.6",
-    "fs-extra": "^6.0.0",
     "handlebars": "^4.0.11",
     "hbs": "^4.0.1",
     "helmet": "^3.12.0",
